pyproject.toml
[project]
name = "energy310-notebooks"
version = "0.1.0"
description = "Course materials for graduate-level energy systems analysis using PyPSA and modern software development practices"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Energy 310 Teaching Team"}
]
license = {text = "MIT"}

dependencies = [
    # Core scientific computing
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "scipy>=1.7.0",
    
    # Jupyter environment
    "jupyter>=1.0.0",
    "jupyterlab>=3.0.0",
    "ipywidgets>=7.6.0",
    
    # PyPSA and energy modeling
    "pypsa>=0.21.0",
    
    # Optimization
    "pulp>=2.5.0",
    
    # Network analysis
    "networkx>=2.6.0",
    
    # Additional utilities
    "requests>=2.25.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
geographic = [
    # Geographic and geospatial packages
    "geopandas>=0.10.0",
    "cartopy>=0.20.0",
    "rasterio>=1.2.0",
    "xarray>=0.19.0",
    "netcdf4",
]
advanced = [
    # Advanced packages that might have complex dependencies
    "atlite>=0.2.4",
    "powerplantmatching>=0.4.8",
    "snakemake>=7.0.0",
    "plotly>=5.0.0",
    "hvplot>=0.7.0",
    "gitpython>=3.1.0",
]

[project.urls]
Homepage = "https://github.com/ktehranchi/energy310_notebooks"
Repository = "https://github.com/ktehranchi/energy310_notebooks"

# No build system since this is not a package

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
]